+++ cmake/format.cmake
@@ -32,7 +32,7 @@ file(
 list(FILTER SRC EXCLUDE REGEX "silkworm/interfaces/")
 list(FILTER SRC EXCLUDE REGEX "silkworm/core/chain/genesis_[a-z_]+.cpp\$")
 list(FILTER SRC EXCLUDE REGEX "silkworm/core/chain/dao.hpp$")
-list(FILTER SRC EXCLUDE REGEX "silkworm/node/common/preverified_hashes_[a-z]+.cpp\$")
 list(FILTER SRC EXCLUDE REGEX "silkworm/rpc/json_rpc/specification.cpp\$")
+list(FILTER SRC EXCLUDE REGEX "silkworm/sync/internals/preverified_hashes/preverified_hashes_[a-z]+.cpp\$")
 
 execute_process(COMMAND ${CLANG_FORMAT} -style=file -i ${SRC})
